refactor: fix for bare import
