+++ src/lib/reillyDOM/index.js
@@ -1,0 +1,1 @@
+export { default } from './ReillyDOM';
